Fix target files.
