+++ internal/codeintel/uploads/internal/background/mocks_test.go
@@ -13,19 +13,16 @@ import (
 	"time"
 
 	regexp "github.com/grafana/regexp"
-	sqlf "github.com/keegancsmith/sqlf"
 	log "github.com/sourcegraph/log"
 	api "github.com/sourcegraph/sourcegraph/internal/api"
 	authz "github.com/sourcegraph/sourcegraph/internal/authz"
 	types "github.com/sourcegraph/sourcegraph/internal/codeintel/shared/types"
 	shared "github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/shared"
-	basestore "github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	gitserver "github.com/sourcegraph/sourcegraph/internal/gitserver"
 	gitdomain "github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	protocol "github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
 	observation "github.com/sourcegraph/sourcegraph/internal/observation"
 	uploadstore "github.com/sourcegraph/sourcegraph/internal/uploadstore"
-	workerutil "github.com/sourcegraph/sourcegraph/internal/workerutil"
 	store "github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker/store"
 )
 
@@ -4511,1796 +4508,4 @@ func (c UploadServiceUpdateSourcedCommitsFuncCall) Args() []interface{} {
 // invocation.
 func (c UploadServiceUpdateSourcedCommitsFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0, c.Result1}
-}
-
-// MockWorkerStore is a mock implementation of the Store interface (from the
-// package
-// github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker/store)
-// used for unit testing.
-type MockWorkerStore struct {
-	// AddExecutionLogEntryFunc is an instance of a mock function object
-	// controlling the behavior of the method AddExecutionLogEntry.
-	AddExecutionLogEntryFunc *WorkerStoreAddExecutionLogEntryFunc
-	// CanceledJobsFunc is an instance of a mock function object controlling
-	// the behavior of the method CanceledJobs.
-	CanceledJobsFunc *WorkerStoreCanceledJobsFunc
-	// DequeueFunc is an instance of a mock function object controlling the
-	// behavior of the method Dequeue.
-	DequeueFunc *WorkerStoreDequeueFunc
-	// HandleFunc is an instance of a mock function object controlling the
-	// behavior of the method Handle.
-	HandleFunc *WorkerStoreHandleFunc
-	// HeartbeatFunc is an instance of a mock function object controlling
-	// the behavior of the method Heartbeat.
-	HeartbeatFunc *WorkerStoreHeartbeatFunc
-	// MarkCompleteFunc is an instance of a mock function object controlling
-	// the behavior of the method MarkComplete.
-	MarkCompleteFunc *WorkerStoreMarkCompleteFunc
-	// MarkErroredFunc is an instance of a mock function object controlling
-	// the behavior of the method MarkErrored.
-	MarkErroredFunc *WorkerStoreMarkErroredFunc
-	// MarkFailedFunc is an instance of a mock function object controlling
-	// the behavior of the method MarkFailed.
-	MarkFailedFunc *WorkerStoreMarkFailedFunc
-	// MaxDurationInQueueFunc is an instance of a mock function object
-	// controlling the behavior of the method MaxDurationInQueue.
-	MaxDurationInQueueFunc *WorkerStoreMaxDurationInQueueFunc
-	// QueuedCountFunc is an instance of a mock function object controlling
-	// the behavior of the method QueuedCount.
-	QueuedCountFunc *WorkerStoreQueuedCountFunc
-	// RequeueFunc is an instance of a mock function object controlling the
-	// behavior of the method Requeue.
-	RequeueFunc *WorkerStoreRequeueFunc
-	// ResetStalledFunc is an instance of a mock function object controlling
-	// the behavior of the method ResetStalled.
-	ResetStalledFunc *WorkerStoreResetStalledFunc
-	// UpdateExecutionLogEntryFunc is an instance of a mock function object
-	// controlling the behavior of the method UpdateExecutionLogEntry.
-	UpdateExecutionLogEntryFunc *WorkerStoreUpdateExecutionLogEntryFunc
-	// WithFunc is an instance of a mock function object controlling the
-	// behavior of the method With.
-	WithFunc *WorkerStoreWithFunc
-}
-
-// NewMockWorkerStore creates a new mock of the Store interface. All methods
-// return zero values for all results, unless overwritten.
-func NewMockWorkerStore() *MockWorkerStore {
-	return &MockWorkerStore{
-		AddExecutionLogEntryFunc: &WorkerStoreAddExecutionLogEntryFunc{
-			defaultHook: func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (r0 int, r1 error) {
-				return
-			},
-		},
-		CanceledJobsFunc: &WorkerStoreCanceledJobsFunc{
-			defaultHook: func(context.Context, []int, store.CanceledJobsOptions) (r0 []int, r1 error) {
-				return
-			},
-		},
-		DequeueFunc: &WorkerStoreDequeueFunc{
-			defaultHook: func(context.Context, string, []*sqlf.Query) (r0 workerutil.Record, r1 bool, r2 error) {
-				return
-			},
-		},
-		HandleFunc: &WorkerStoreHandleFunc{
-			defaultHook: func() (r0 basestore.TransactableHandle) {
-				return
-			},
-		},
-		HeartbeatFunc: &WorkerStoreHeartbeatFunc{
-			defaultHook: func(context.Context, []int, store.HeartbeatOptions) (r0 []int, r1 error) {
-				return
-			},
-		},
-		MarkCompleteFunc: &WorkerStoreMarkCompleteFunc{
-			defaultHook: func(context.Context, int, store.MarkFinalOptions) (r0 bool, r1 error) {
-				return
-			},
-		},
-		MarkErroredFunc: &WorkerStoreMarkErroredFunc{
-			defaultHook: func(context.Context, int, string, store.MarkFinalOptions) (r0 bool, r1 error) {
-				return
-			},
-		},
-		MarkFailedFunc: &WorkerStoreMarkFailedFunc{
-			defaultHook: func(context.Context, int, string, store.MarkFinalOptions) (r0 bool, r1 error) {
-				return
-			},
-		},
-		MaxDurationInQueueFunc: &WorkerStoreMaxDurationInQueueFunc{
-			defaultHook: func(context.Context) (r0 time.Duration, r1 error) {
-				return
-			},
-		},
-		QueuedCountFunc: &WorkerStoreQueuedCountFunc{
-			defaultHook: func(context.Context, bool) (r0 int, r1 error) {
-				return
-			},
-		},
-		RequeueFunc: &WorkerStoreRequeueFunc{
-			defaultHook: func(context.Context, int, time.Time) (r0 error) {
-				return
-			},
-		},
-		ResetStalledFunc: &WorkerStoreResetStalledFunc{
-			defaultHook: func(context.Context) (r0 map[int]time.Duration, r1 map[int]time.Duration, r2 error) {
-				return
-			},
-		},
-		UpdateExecutionLogEntryFunc: &WorkerStoreUpdateExecutionLogEntryFunc{
-			defaultHook: func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (r0 error) {
-				return
-			},
-		},
-		WithFunc: &WorkerStoreWithFunc{
-			defaultHook: func(basestore.ShareableStore) (r0 store.Store) {
-				return
-			},
-		},
-	}
-}
-
-// NewStrictMockWorkerStore creates a new mock of the Store interface. All
-// methods panic on invocation, unless overwritten.
-func NewStrictMockWorkerStore() *MockWorkerStore {
-	return &MockWorkerStore{
-		AddExecutionLogEntryFunc: &WorkerStoreAddExecutionLogEntryFunc{
-			defaultHook: func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error) {
-				panic("unexpected invocation of MockWorkerStore.AddExecutionLogEntry")
-			},
-		},
-		CanceledJobsFunc: &WorkerStoreCanceledJobsFunc{
-			defaultHook: func(context.Context, []int, store.CanceledJobsOptions) ([]int, error) {
-				panic("unexpected invocation of MockWorkerStore.CanceledJobs")
-			},
-		},
-		DequeueFunc: &WorkerStoreDequeueFunc{
-			defaultHook: func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error) {
-				panic("unexpected invocation of MockWorkerStore.Dequeue")
-			},
-		},
-		HandleFunc: &WorkerStoreHandleFunc{
-			defaultHook: func() basestore.TransactableHandle {
-				panic("unexpected invocation of MockWorkerStore.Handle")
-			},
-		},
-		HeartbeatFunc: &WorkerStoreHeartbeatFunc{
-			defaultHook: func(context.Context, []int, store.HeartbeatOptions) ([]int, error) {
-				panic("unexpected invocation of MockWorkerStore.Heartbeat")
-			},
-		},
-		MarkCompleteFunc: &WorkerStoreMarkCompleteFunc{
-			defaultHook: func(context.Context, int, store.MarkFinalOptions) (bool, error) {
-				panic("unexpected invocation of MockWorkerStore.MarkComplete")
-			},
-		},
-		MarkErroredFunc: &WorkerStoreMarkErroredFunc{
-			defaultHook: func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-				panic("unexpected invocation of MockWorkerStore.MarkErrored")
-			},
-		},
-		MarkFailedFunc: &WorkerStoreMarkFailedFunc{
-			defaultHook: func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-				panic("unexpected invocation of MockWorkerStore.MarkFailed")
-			},
-		},
-		MaxDurationInQueueFunc: &WorkerStoreMaxDurationInQueueFunc{
-			defaultHook: func(context.Context) (time.Duration, error) {
-				panic("unexpected invocation of MockWorkerStore.MaxDurationInQueue")
-			},
-		},
-		QueuedCountFunc: &WorkerStoreQueuedCountFunc{
-			defaultHook: func(context.Context, bool) (int, error) {
-				panic("unexpected invocation of MockWorkerStore.QueuedCount")
-			},
-		},
-		RequeueFunc: &WorkerStoreRequeueFunc{
-			defaultHook: func(context.Context, int, time.Time) error {
-				panic("unexpected invocation of MockWorkerStore.Requeue")
-			},
-		},
-		ResetStalledFunc: &WorkerStoreResetStalledFunc{
-			defaultHook: func(context.Context) (map[int]time.Duration, map[int]time.Duration, error) {
-				panic("unexpected invocation of MockWorkerStore.ResetStalled")
-			},
-		},
-		UpdateExecutionLogEntryFunc: &WorkerStoreUpdateExecutionLogEntryFunc{
-			defaultHook: func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error {
-				panic("unexpected invocation of MockWorkerStore.UpdateExecutionLogEntry")
-			},
-		},
-		WithFunc: &WorkerStoreWithFunc{
-			defaultHook: func(basestore.ShareableStore) store.Store {
-				panic("unexpected invocation of MockWorkerStore.With")
-			},
-		},
-	}
-}
-
-// NewMockWorkerStoreFrom creates a new mock of the MockWorkerStore
-// interface. All methods delegate to the given implementation, unless
-// overwritten.
-func NewMockWorkerStoreFrom(i store.Store) *MockWorkerStore {
-	return &MockWorkerStore{
-		AddExecutionLogEntryFunc: &WorkerStoreAddExecutionLogEntryFunc{
-			defaultHook: i.AddExecutionLogEntry,
-		},
-		CanceledJobsFunc: &WorkerStoreCanceledJobsFunc{
-			defaultHook: i.CanceledJobs,
-		},
-		DequeueFunc: &WorkerStoreDequeueFunc{
-			defaultHook: i.Dequeue,
-		},
-		HandleFunc: &WorkerStoreHandleFunc{
-			defaultHook: i.Handle,
-		},
-		HeartbeatFunc: &WorkerStoreHeartbeatFunc{
-			defaultHook: i.Heartbeat,
-		},
-		MarkCompleteFunc: &WorkerStoreMarkCompleteFunc{
-			defaultHook: i.MarkComplete,
-		},
-		MarkErroredFunc: &WorkerStoreMarkErroredFunc{
-			defaultHook: i.MarkErrored,
-		},
-		MarkFailedFunc: &WorkerStoreMarkFailedFunc{
-			defaultHook: i.MarkFailed,
-		},
-		MaxDurationInQueueFunc: &WorkerStoreMaxDurationInQueueFunc{
-			defaultHook: i.MaxDurationInQueue,
-		},
-		QueuedCountFunc: &WorkerStoreQueuedCountFunc{
-			defaultHook: i.QueuedCount,
-		},
-		RequeueFunc: &WorkerStoreRequeueFunc{
-			defaultHook: i.Requeue,
-		},
-		ResetStalledFunc: &WorkerStoreResetStalledFunc{
-			defaultHook: i.ResetStalled,
-		},
-		UpdateExecutionLogEntryFunc: &WorkerStoreUpdateExecutionLogEntryFunc{
-			defaultHook: i.UpdateExecutionLogEntry,
-		},
-		WithFunc: &WorkerStoreWithFunc{
-			defaultHook: i.With,
-		},
-	}
-}
-
-// WorkerStoreAddExecutionLogEntryFunc describes the behavior when the
-// AddExecutionLogEntry method of the parent MockWorkerStore instance is
-// invoked.
-type WorkerStoreAddExecutionLogEntryFunc struct {
-	defaultHook func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error)
-	hooks       []func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error)
-	history     []WorkerStoreAddExecutionLogEntryFuncCall
-	mutex       sync.Mutex
-}
-
-// AddExecutionLogEntry delegates to the next hook function in the queue and
-// stores the parameter and result values of this invocation.
-func (m *MockWorkerStore) AddExecutionLogEntry(v0 context.Context, v1 int, v2 workerutil.ExecutionLogEntry, v3 store.ExecutionLogEntryOptions) (int, error) {
-	r0, r1 := m.AddExecutionLogEntryFunc.nextHook()(v0, v1, v2, v3)
-	m.AddExecutionLogEntryFunc.appendCall(WorkerStoreAddExecutionLogEntryFuncCall{v0, v1, v2, v3, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the AddExecutionLogEntry
-// method of the parent MockWorkerStore instance is invoked and the hook
-// queue is empty.
-func (f *WorkerStoreAddExecutionLogEntryFunc) SetDefaultHook(hook func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// AddExecutionLogEntry method of the parent MockWorkerStore instance
-// invokes the hook at the front of the queue and discards it. After the
-// queue is empty, the default hook function is invoked for any future
-// action.
-func (f *WorkerStoreAddExecutionLogEntryFunc) PushHook(hook func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreAddExecutionLogEntryFunc) SetDefaultReturn(r0 int, r1 error) {
-	f.SetDefaultHook(func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreAddExecutionLogEntryFunc) PushReturn(r0 int, r1 error) {
-	f.PushHook(func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreAddExecutionLogEntryFunc) nextHook() func(context.Context, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) (int, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreAddExecutionLogEntryFunc) appendCall(r0 WorkerStoreAddExecutionLogEntryFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreAddExecutionLogEntryFuncCall
-// objects describing the invocations of this function.
-func (f *WorkerStoreAddExecutionLogEntryFunc) History() []WorkerStoreAddExecutionLogEntryFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreAddExecutionLogEntryFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreAddExecutionLogEntryFuncCall is an object that describes an
-// invocation of method AddExecutionLogEntry on an instance of
-// MockWorkerStore.
-type WorkerStoreAddExecutionLogEntryFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 workerutil.ExecutionLogEntry
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 store.ExecutionLogEntryOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 int
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreAddExecutionLogEntryFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreAddExecutionLogEntryFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreCanceledJobsFunc describes the behavior when the CanceledJobs
-// method of the parent MockWorkerStore instance is invoked.
-type WorkerStoreCanceledJobsFunc struct {
-	defaultHook func(context.Context, []int, store.CanceledJobsOptions) ([]int, error)
-	hooks       []func(context.Context, []int, store.CanceledJobsOptions) ([]int, error)
-	history     []WorkerStoreCanceledJobsFuncCall
-	mutex       sync.Mutex
-}
-
-// CanceledJobs delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockWorkerStore) CanceledJobs(v0 context.Context, v1 []int, v2 store.CanceledJobsOptions) ([]int, error) {
-	r0, r1 := m.CanceledJobsFunc.nextHook()(v0, v1, v2)
-	m.CanceledJobsFunc.appendCall(WorkerStoreCanceledJobsFuncCall{v0, v1, v2, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the CanceledJobs method
-// of the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreCanceledJobsFunc) SetDefaultHook(hook func(context.Context, []int, store.CanceledJobsOptions) ([]int, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// CanceledJobs method of the parent MockWorkerStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *WorkerStoreCanceledJobsFunc) PushHook(hook func(context.Context, []int, store.CanceledJobsOptions) ([]int, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreCanceledJobsFunc) SetDefaultReturn(r0 []int, r1 error) {
-	f.SetDefaultHook(func(context.Context, []int, store.CanceledJobsOptions) ([]int, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreCanceledJobsFunc) PushReturn(r0 []int, r1 error) {
-	f.PushHook(func(context.Context, []int, store.CanceledJobsOptions) ([]int, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreCanceledJobsFunc) nextHook() func(context.Context, []int, store.CanceledJobsOptions) ([]int, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreCanceledJobsFunc) appendCall(r0 WorkerStoreCanceledJobsFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreCanceledJobsFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreCanceledJobsFunc) History() []WorkerStoreCanceledJobsFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreCanceledJobsFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreCanceledJobsFuncCall is an object that describes an invocation
-// of method CanceledJobs on an instance of MockWorkerStore.
-type WorkerStoreCanceledJobsFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 []int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 store.CanceledJobsOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 []int
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreCanceledJobsFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreCanceledJobsFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreDequeueFunc describes the behavior when the Dequeue method of
-// the parent MockWorkerStore instance is invoked.
-type WorkerStoreDequeueFunc struct {
-	defaultHook func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error)
-	hooks       []func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error)
-	history     []WorkerStoreDequeueFuncCall
-	mutex       sync.Mutex
-}
-
-// Dequeue delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockWorkerStore) Dequeue(v0 context.Context, v1 string, v2 []*sqlf.Query) (workerutil.Record, bool, error) {
-	r0, r1, r2 := m.DequeueFunc.nextHook()(v0, v1, v2)
-	m.DequeueFunc.appendCall(WorkerStoreDequeueFuncCall{v0, v1, v2, r0, r1, r2})
-	return r0, r1, r2
-}
-
-// SetDefaultHook sets function that is called when the Dequeue method of
-// the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreDequeueFunc) SetDefaultHook(hook func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Dequeue method of the parent MockWorkerStore instance invokes the hook at
-// the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *WorkerStoreDequeueFunc) PushHook(hook func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreDequeueFunc) SetDefaultReturn(r0 workerutil.Record, r1 bool, r2 error) {
-	f.SetDefaultHook(func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error) {
-		return r0, r1, r2
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreDequeueFunc) PushReturn(r0 workerutil.Record, r1 bool, r2 error) {
-	f.PushHook(func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error) {
-		return r0, r1, r2
-	})
-}
-
-func (f *WorkerStoreDequeueFunc) nextHook() func(context.Context, string, []*sqlf.Query) (workerutil.Record, bool, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreDequeueFunc) appendCall(r0 WorkerStoreDequeueFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreDequeueFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreDequeueFunc) History() []WorkerStoreDequeueFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreDequeueFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreDequeueFuncCall is an object that describes an invocation of
-// method Dequeue on an instance of MockWorkerStore.
-type WorkerStoreDequeueFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 string
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 []*sqlf.Query
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 workerutil.Record
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 bool
-	// Result2 is the value of the 3rd result returned from this method
-	// invocation.
-	Result2 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreDequeueFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreDequeueFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1, c.Result2}
-}
-
-// WorkerStoreHandleFunc describes the behavior when the Handle method of
-// the parent MockWorkerStore instance is invoked.
-type WorkerStoreHandleFunc struct {
-	defaultHook func() basestore.TransactableHandle
-	hooks       []func() basestore.TransactableHandle
-	history     []WorkerStoreHandleFuncCall
-	mutex       sync.Mutex
-}
-
-// Handle delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockWorkerStore) Handle() basestore.TransactableHandle {
-	r0 := m.HandleFunc.nextHook()()
-	m.HandleFunc.appendCall(WorkerStoreHandleFuncCall{r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the Handle method of the
-// parent MockWorkerStore instance is invoked and the hook queue is empty.
-func (f *WorkerStoreHandleFunc) SetDefaultHook(hook func() basestore.TransactableHandle) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Handle method of the parent MockWorkerStore instance invokes the hook at
-// the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *WorkerStoreHandleFunc) PushHook(hook func() basestore.TransactableHandle) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreHandleFunc) SetDefaultReturn(r0 basestore.TransactableHandle) {
-	f.SetDefaultHook(func() basestore.TransactableHandle {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreHandleFunc) PushReturn(r0 basestore.TransactableHandle) {
-	f.PushHook(func() basestore.TransactableHandle {
-		return r0
-	})
-}
-
-func (f *WorkerStoreHandleFunc) nextHook() func() basestore.TransactableHandle {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreHandleFunc) appendCall(r0 WorkerStoreHandleFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreHandleFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreHandleFunc) History() []WorkerStoreHandleFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreHandleFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreHandleFuncCall is an object that describes an invocation of
-// method Handle on an instance of MockWorkerStore.
-type WorkerStoreHandleFuncCall struct {
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 basestore.TransactableHandle
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreHandleFuncCall) Args() []interface{} {
-	return []interface{}{}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreHandleFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// WorkerStoreHeartbeatFunc describes the behavior when the Heartbeat method
-// of the parent MockWorkerStore instance is invoked.
-type WorkerStoreHeartbeatFunc struct {
-	defaultHook func(context.Context, []int, store.HeartbeatOptions) ([]int, error)
-	hooks       []func(context.Context, []int, store.HeartbeatOptions) ([]int, error)
-	history     []WorkerStoreHeartbeatFuncCall
-	mutex       sync.Mutex
-}
-
-// Heartbeat delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockWorkerStore) Heartbeat(v0 context.Context, v1 []int, v2 store.HeartbeatOptions) ([]int, error) {
-	r0, r1 := m.HeartbeatFunc.nextHook()(v0, v1, v2)
-	m.HeartbeatFunc.appendCall(WorkerStoreHeartbeatFuncCall{v0, v1, v2, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the Heartbeat method of
-// the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreHeartbeatFunc) SetDefaultHook(hook func(context.Context, []int, store.HeartbeatOptions) ([]int, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Heartbeat method of the parent MockWorkerStore instance invokes the hook
-// at the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *WorkerStoreHeartbeatFunc) PushHook(hook func(context.Context, []int, store.HeartbeatOptions) ([]int, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreHeartbeatFunc) SetDefaultReturn(r0 []int, r1 error) {
-	f.SetDefaultHook(func(context.Context, []int, store.HeartbeatOptions) ([]int, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreHeartbeatFunc) PushReturn(r0 []int, r1 error) {
-	f.PushHook(func(context.Context, []int, store.HeartbeatOptions) ([]int, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreHeartbeatFunc) nextHook() func(context.Context, []int, store.HeartbeatOptions) ([]int, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreHeartbeatFunc) appendCall(r0 WorkerStoreHeartbeatFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreHeartbeatFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreHeartbeatFunc) History() []WorkerStoreHeartbeatFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreHeartbeatFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreHeartbeatFuncCall is an object that describes an invocation of
-// method Heartbeat on an instance of MockWorkerStore.
-type WorkerStoreHeartbeatFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 []int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 store.HeartbeatOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 []int
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreHeartbeatFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreHeartbeatFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreMarkCompleteFunc describes the behavior when the MarkComplete
-// method of the parent MockWorkerStore instance is invoked.
-type WorkerStoreMarkCompleteFunc struct {
-	defaultHook func(context.Context, int, store.MarkFinalOptions) (bool, error)
-	hooks       []func(context.Context, int, store.MarkFinalOptions) (bool, error)
-	history     []WorkerStoreMarkCompleteFuncCall
-	mutex       sync.Mutex
-}
-
-// MarkComplete delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockWorkerStore) MarkComplete(v0 context.Context, v1 int, v2 store.MarkFinalOptions) (bool, error) {
-	r0, r1 := m.MarkCompleteFunc.nextHook()(v0, v1, v2)
-	m.MarkCompleteFunc.appendCall(WorkerStoreMarkCompleteFuncCall{v0, v1, v2, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the MarkComplete method
-// of the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreMarkCompleteFunc) SetDefaultHook(hook func(context.Context, int, store.MarkFinalOptions) (bool, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// MarkComplete method of the parent MockWorkerStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *WorkerStoreMarkCompleteFunc) PushHook(hook func(context.Context, int, store.MarkFinalOptions) (bool, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreMarkCompleteFunc) SetDefaultReturn(r0 bool, r1 error) {
-	f.SetDefaultHook(func(context.Context, int, store.MarkFinalOptions) (bool, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreMarkCompleteFunc) PushReturn(r0 bool, r1 error) {
-	f.PushHook(func(context.Context, int, store.MarkFinalOptions) (bool, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreMarkCompleteFunc) nextHook() func(context.Context, int, store.MarkFinalOptions) (bool, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreMarkCompleteFunc) appendCall(r0 WorkerStoreMarkCompleteFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreMarkCompleteFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreMarkCompleteFunc) History() []WorkerStoreMarkCompleteFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreMarkCompleteFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreMarkCompleteFuncCall is an object that describes an invocation
-// of method MarkComplete on an instance of MockWorkerStore.
-type WorkerStoreMarkCompleteFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 store.MarkFinalOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 bool
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreMarkCompleteFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreMarkCompleteFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreMarkErroredFunc describes the behavior when the MarkErrored
-// method of the parent MockWorkerStore instance is invoked.
-type WorkerStoreMarkErroredFunc struct {
-	defaultHook func(context.Context, int, string, store.MarkFinalOptions) (bool, error)
-	hooks       []func(context.Context, int, string, store.MarkFinalOptions) (bool, error)
-	history     []WorkerStoreMarkErroredFuncCall
-	mutex       sync.Mutex
-}
-
-// MarkErrored delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockWorkerStore) MarkErrored(v0 context.Context, v1 int, v2 string, v3 store.MarkFinalOptions) (bool, error) {
-	r0, r1 := m.MarkErroredFunc.nextHook()(v0, v1, v2, v3)
-	m.MarkErroredFunc.appendCall(WorkerStoreMarkErroredFuncCall{v0, v1, v2, v3, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the MarkErrored method
-// of the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreMarkErroredFunc) SetDefaultHook(hook func(context.Context, int, string, store.MarkFinalOptions) (bool, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// MarkErrored method of the parent MockWorkerStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *WorkerStoreMarkErroredFunc) PushHook(hook func(context.Context, int, string, store.MarkFinalOptions) (bool, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreMarkErroredFunc) SetDefaultReturn(r0 bool, r1 error) {
-	f.SetDefaultHook(func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreMarkErroredFunc) PushReturn(r0 bool, r1 error) {
-	f.PushHook(func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreMarkErroredFunc) nextHook() func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreMarkErroredFunc) appendCall(r0 WorkerStoreMarkErroredFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreMarkErroredFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreMarkErroredFunc) History() []WorkerStoreMarkErroredFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreMarkErroredFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreMarkErroredFuncCall is an object that describes an invocation
-// of method MarkErrored on an instance of MockWorkerStore.
-type WorkerStoreMarkErroredFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 string
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 store.MarkFinalOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 bool
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreMarkErroredFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreMarkErroredFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreMarkFailedFunc describes the behavior when the MarkFailed
-// method of the parent MockWorkerStore instance is invoked.
-type WorkerStoreMarkFailedFunc struct {
-	defaultHook func(context.Context, int, string, store.MarkFinalOptions) (bool, error)
-	hooks       []func(context.Context, int, string, store.MarkFinalOptions) (bool, error)
-	history     []WorkerStoreMarkFailedFuncCall
-	mutex       sync.Mutex
-}
-
-// MarkFailed delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockWorkerStore) MarkFailed(v0 context.Context, v1 int, v2 string, v3 store.MarkFinalOptions) (bool, error) {
-	r0, r1 := m.MarkFailedFunc.nextHook()(v0, v1, v2, v3)
-	m.MarkFailedFunc.appendCall(WorkerStoreMarkFailedFuncCall{v0, v1, v2, v3, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the MarkFailed method of
-// the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreMarkFailedFunc) SetDefaultHook(hook func(context.Context, int, string, store.MarkFinalOptions) (bool, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// MarkFailed method of the parent MockWorkerStore instance invokes the hook
-// at the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *WorkerStoreMarkFailedFunc) PushHook(hook func(context.Context, int, string, store.MarkFinalOptions) (bool, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreMarkFailedFunc) SetDefaultReturn(r0 bool, r1 error) {
-	f.SetDefaultHook(func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreMarkFailedFunc) PushReturn(r0 bool, r1 error) {
-	f.PushHook(func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreMarkFailedFunc) nextHook() func(context.Context, int, string, store.MarkFinalOptions) (bool, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreMarkFailedFunc) appendCall(r0 WorkerStoreMarkFailedFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreMarkFailedFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreMarkFailedFunc) History() []WorkerStoreMarkFailedFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreMarkFailedFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreMarkFailedFuncCall is an object that describes an invocation
-// of method MarkFailed on an instance of MockWorkerStore.
-type WorkerStoreMarkFailedFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 string
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 store.MarkFinalOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 bool
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreMarkFailedFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreMarkFailedFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreMaxDurationInQueueFunc describes the behavior when the
-// MaxDurationInQueue method of the parent MockWorkerStore instance is
-// invoked.
-type WorkerStoreMaxDurationInQueueFunc struct {
-	defaultHook func(context.Context) (time.Duration, error)
-	hooks       []func(context.Context) (time.Duration, error)
-	history     []WorkerStoreMaxDurationInQueueFuncCall
-	mutex       sync.Mutex
-}
-
-// MaxDurationInQueue delegates to the next hook function in the queue and
-// stores the parameter and result values of this invocation.
-func (m *MockWorkerStore) MaxDurationInQueue(v0 context.Context) (time.Duration, error) {
-	r0, r1 := m.MaxDurationInQueueFunc.nextHook()(v0)
-	m.MaxDurationInQueueFunc.appendCall(WorkerStoreMaxDurationInQueueFuncCall{v0, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the MaxDurationInQueue
-// method of the parent MockWorkerStore instance is invoked and the hook
-// queue is empty.
-func (f *WorkerStoreMaxDurationInQueueFunc) SetDefaultHook(hook func(context.Context) (time.Duration, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// MaxDurationInQueue method of the parent MockWorkerStore instance invokes
-// the hook at the front of the queue and discards it. After the queue is
-// empty, the default hook function is invoked for any future action.
-func (f *WorkerStoreMaxDurationInQueueFunc) PushHook(hook func(context.Context) (time.Duration, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreMaxDurationInQueueFunc) SetDefaultReturn(r0 time.Duration, r1 error) {
-	f.SetDefaultHook(func(context.Context) (time.Duration, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreMaxDurationInQueueFunc) PushReturn(r0 time.Duration, r1 error) {
-	f.PushHook(func(context.Context) (time.Duration, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreMaxDurationInQueueFunc) nextHook() func(context.Context) (time.Duration, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreMaxDurationInQueueFunc) appendCall(r0 WorkerStoreMaxDurationInQueueFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreMaxDurationInQueueFuncCall
-// objects describing the invocations of this function.
-func (f *WorkerStoreMaxDurationInQueueFunc) History() []WorkerStoreMaxDurationInQueueFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreMaxDurationInQueueFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreMaxDurationInQueueFuncCall is an object that describes an
-// invocation of method MaxDurationInQueue on an instance of
-// MockWorkerStore.
-type WorkerStoreMaxDurationInQueueFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 time.Duration
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreMaxDurationInQueueFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreMaxDurationInQueueFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreQueuedCountFunc describes the behavior when the QueuedCount
-// method of the parent MockWorkerStore instance is invoked.
-type WorkerStoreQueuedCountFunc struct {
-	defaultHook func(context.Context, bool) (int, error)
-	hooks       []func(context.Context, bool) (int, error)
-	history     []WorkerStoreQueuedCountFuncCall
-	mutex       sync.Mutex
-}
-
-// QueuedCount delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockWorkerStore) QueuedCount(v0 context.Context, v1 bool) (int, error) {
-	r0, r1 := m.QueuedCountFunc.nextHook()(v0, v1)
-	m.QueuedCountFunc.appendCall(WorkerStoreQueuedCountFuncCall{v0, v1, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the QueuedCount method
-// of the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreQueuedCountFunc) SetDefaultHook(hook func(context.Context, bool) (int, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// QueuedCount method of the parent MockWorkerStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *WorkerStoreQueuedCountFunc) PushHook(hook func(context.Context, bool) (int, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreQueuedCountFunc) SetDefaultReturn(r0 int, r1 error) {
-	f.SetDefaultHook(func(context.Context, bool) (int, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreQueuedCountFunc) PushReturn(r0 int, r1 error) {
-	f.PushHook(func(context.Context, bool) (int, error) {
-		return r0, r1
-	})
-}
-
-func (f *WorkerStoreQueuedCountFunc) nextHook() func(context.Context, bool) (int, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreQueuedCountFunc) appendCall(r0 WorkerStoreQueuedCountFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreQueuedCountFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreQueuedCountFunc) History() []WorkerStoreQueuedCountFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreQueuedCountFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreQueuedCountFuncCall is an object that describes an invocation
-// of method QueuedCount on an instance of MockWorkerStore.
-type WorkerStoreQueuedCountFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 bool
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 int
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreQueuedCountFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreQueuedCountFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// WorkerStoreRequeueFunc describes the behavior when the Requeue method of
-// the parent MockWorkerStore instance is invoked.
-type WorkerStoreRequeueFunc struct {
-	defaultHook func(context.Context, int, time.Time) error
-	hooks       []func(context.Context, int, time.Time) error
-	history     []WorkerStoreRequeueFuncCall
-	mutex       sync.Mutex
-}
-
-// Requeue delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockWorkerStore) Requeue(v0 context.Context, v1 int, v2 time.Time) error {
-	r0 := m.RequeueFunc.nextHook()(v0, v1, v2)
-	m.RequeueFunc.appendCall(WorkerStoreRequeueFuncCall{v0, v1, v2, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the Requeue method of
-// the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreRequeueFunc) SetDefaultHook(hook func(context.Context, int, time.Time) error) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Requeue method of the parent MockWorkerStore instance invokes the hook at
-// the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *WorkerStoreRequeueFunc) PushHook(hook func(context.Context, int, time.Time) error) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreRequeueFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(context.Context, int, time.Time) error {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreRequeueFunc) PushReturn(r0 error) {
-	f.PushHook(func(context.Context, int, time.Time) error {
-		return r0
-	})
-}
-
-func (f *WorkerStoreRequeueFunc) nextHook() func(context.Context, int, time.Time) error {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreRequeueFunc) appendCall(r0 WorkerStoreRequeueFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreRequeueFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreRequeueFunc) History() []WorkerStoreRequeueFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreRequeueFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreRequeueFuncCall is an object that describes an invocation of
-// method Requeue on an instance of MockWorkerStore.
-type WorkerStoreRequeueFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 time.Time
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreRequeueFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreRequeueFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// WorkerStoreResetStalledFunc describes the behavior when the ResetStalled
-// method of the parent MockWorkerStore instance is invoked.
-type WorkerStoreResetStalledFunc struct {
-	defaultHook func(context.Context) (map[int]time.Duration, map[int]time.Duration, error)
-	hooks       []func(context.Context) (map[int]time.Duration, map[int]time.Duration, error)
-	history     []WorkerStoreResetStalledFuncCall
-	mutex       sync.Mutex
-}
-
-// ResetStalled delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockWorkerStore) ResetStalled(v0 context.Context) (map[int]time.Duration, map[int]time.Duration, error) {
-	r0, r1, r2 := m.ResetStalledFunc.nextHook()(v0)
-	m.ResetStalledFunc.appendCall(WorkerStoreResetStalledFuncCall{v0, r0, r1, r2})
-	return r0, r1, r2
-}
-
-// SetDefaultHook sets function that is called when the ResetStalled method
-// of the parent MockWorkerStore instance is invoked and the hook queue is
-// empty.
-func (f *WorkerStoreResetStalledFunc) SetDefaultHook(hook func(context.Context) (map[int]time.Duration, map[int]time.Duration, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// ResetStalled method of the parent MockWorkerStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *WorkerStoreResetStalledFunc) PushHook(hook func(context.Context) (map[int]time.Duration, map[int]time.Duration, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreResetStalledFunc) SetDefaultReturn(r0 map[int]time.Duration, r1 map[int]time.Duration, r2 error) {
-	f.SetDefaultHook(func(context.Context) (map[int]time.Duration, map[int]time.Duration, error) {
-		return r0, r1, r2
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreResetStalledFunc) PushReturn(r0 map[int]time.Duration, r1 map[int]time.Duration, r2 error) {
-	f.PushHook(func(context.Context) (map[int]time.Duration, map[int]time.Duration, error) {
-		return r0, r1, r2
-	})
-}
-
-func (f *WorkerStoreResetStalledFunc) nextHook() func(context.Context) (map[int]time.Duration, map[int]time.Duration, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreResetStalledFunc) appendCall(r0 WorkerStoreResetStalledFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreResetStalledFuncCall objects
-// describing the invocations of this function.
-func (f *WorkerStoreResetStalledFunc) History() []WorkerStoreResetStalledFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreResetStalledFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreResetStalledFuncCall is an object that describes an invocation
-// of method ResetStalled on an instance of MockWorkerStore.
-type WorkerStoreResetStalledFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 map[int]time.Duration
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 map[int]time.Duration
-	// Result2 is the value of the 3rd result returned from this method
-	// invocation.
-	Result2 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreResetStalledFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreResetStalledFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1, c.Result2}
-}
-
-// WorkerStoreUpdateExecutionLogEntryFunc describes the behavior when the
-// UpdateExecutionLogEntry method of the parent MockWorkerStore instance is
-// invoked.
-type WorkerStoreUpdateExecutionLogEntryFunc struct {
-	defaultHook func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error
-	hooks       []func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error
-	history     []WorkerStoreUpdateExecutionLogEntryFuncCall
-	mutex       sync.Mutex
-}
-
-// UpdateExecutionLogEntry delegates to the next hook function in the queue
-// and stores the parameter and result values of this invocation.
-func (m *MockWorkerStore) UpdateExecutionLogEntry(v0 context.Context, v1 int, v2 int, v3 workerutil.ExecutionLogEntry, v4 store.ExecutionLogEntryOptions) error {
-	r0 := m.UpdateExecutionLogEntryFunc.nextHook()(v0, v1, v2, v3, v4)
-	m.UpdateExecutionLogEntryFunc.appendCall(WorkerStoreUpdateExecutionLogEntryFuncCall{v0, v1, v2, v3, v4, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the
-// UpdateExecutionLogEntry method of the parent MockWorkerStore instance is
-// invoked and the hook queue is empty.
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) SetDefaultHook(hook func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// UpdateExecutionLogEntry method of the parent MockWorkerStore instance
-// invokes the hook at the front of the queue and discards it. After the
-// queue is empty, the default hook function is invoked for any future
-// action.
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) PushHook(hook func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) PushReturn(r0 error) {
-	f.PushHook(func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error {
-		return r0
-	})
-}
-
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) nextHook() func(context.Context, int, int, workerutil.ExecutionLogEntry, store.ExecutionLogEntryOptions) error {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) appendCall(r0 WorkerStoreUpdateExecutionLogEntryFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreUpdateExecutionLogEntryFuncCall
-// objects describing the invocations of this function.
-func (f *WorkerStoreUpdateExecutionLogEntryFunc) History() []WorkerStoreUpdateExecutionLogEntryFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreUpdateExecutionLogEntryFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreUpdateExecutionLogEntryFuncCall is an object that describes an
-// invocation of method UpdateExecutionLogEntry on an instance of
-// MockWorkerStore.
-type WorkerStoreUpdateExecutionLogEntryFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 int
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 workerutil.ExecutionLogEntry
-	// Arg4 is the value of the 5th argument passed to this method
-	// invocation.
-	Arg4 store.ExecutionLogEntryOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreUpdateExecutionLogEntryFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3, c.Arg4}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreUpdateExecutionLogEntryFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// WorkerStoreWithFunc describes the behavior when the With method of the
-// parent MockWorkerStore instance is invoked.
-type WorkerStoreWithFunc struct {
-	defaultHook func(basestore.ShareableStore) store.Store
-	hooks       []func(basestore.ShareableStore) store.Store
-	history     []WorkerStoreWithFuncCall
-	mutex       sync.Mutex
-}
-
-// With delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockWorkerStore) With(v0 basestore.ShareableStore) store.Store {
-	r0 := m.WithFunc.nextHook()(v0)
-	m.WithFunc.appendCall(WorkerStoreWithFuncCall{v0, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the With method of the
-// parent MockWorkerStore instance is invoked and the hook queue is empty.
-func (f *WorkerStoreWithFunc) SetDefaultHook(hook func(basestore.ShareableStore) store.Store) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// With method of the parent MockWorkerStore instance invokes the hook at
-// the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *WorkerStoreWithFunc) PushHook(hook func(basestore.ShareableStore) store.Store) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *WorkerStoreWithFunc) SetDefaultReturn(r0 store.Store) {
-	f.SetDefaultHook(func(basestore.ShareableStore) store.Store {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *WorkerStoreWithFunc) PushReturn(r0 store.Store) {
-	f.PushHook(func(basestore.ShareableStore) store.Store {
-		return r0
-	})
-}
-
-func (f *WorkerStoreWithFunc) nextHook() func(basestore.ShareableStore) store.Store {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *WorkerStoreWithFunc) appendCall(r0 WorkerStoreWithFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of WorkerStoreWithFuncCall objects describing
-// the invocations of this function.
-func (f *WorkerStoreWithFunc) History() []WorkerStoreWithFuncCall {
-	f.mutex.Lock()
-	history := make([]WorkerStoreWithFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// WorkerStoreWithFuncCall is an object that describes an invocation of
-// method With on an instance of MockWorkerStore.
-type WorkerStoreWithFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 basestore.ShareableStore
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 store.Store
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c WorkerStoreWithFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c WorkerStoreWithFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
 }
